campaignion 7.x-2.29 > 7.x-2.32

+++ docs/sites/all/modules/contrib/campaignion/campaignion_wizard/templates/campaignion-wizard-thank-summary.tpl.php
@@ -16,10 +16,12 @@
 ?>
 <?php
 if ($after_submit['type'] == 'node') {
-  echo t('After submitting the last form step your supporters will see the page <a href="@page_href">@page_title</a>.', [
-    '@page_href' => url("node/{$after_submit['node_reference_nid']}"),
-    '@page_title' => $after_submit['node']->title,
-  ]);
+  if ($after_submit['node']) {
+    echo t('After submitting the last form step your supporters will see the page <a href="@page_href">@page_title</a>.', [
+      '@page_href' => url("node/{$after_submit['node_reference_nid']}"),
+      '@page_title' => $after_submit['node']->title,
+    ]);
+  }
 }
 else {
   echo t('After submitting the last form step your supporters will be redirected to one of the following pages:');
@@ -29,10 +31,12 @@ else {
 <?php
 if ($after_confirm) {
   if ($after_confirm['type'] == 'node') {
-    echo t('After clicking on the confirmation link your supporters will see the page <a href="@page_href">@page_title</a>.', [
-      '@page_href' => url("node/{$after_confirm['node_reference_nid']}"),
-      '@page_title' => $after_confirm['node']->title,
-    ]);
+    if ($after_confirm['node']) {
+      echo t('After clicking on the confirmation link your supporters will see the page <a href="@page_href">@page_title</a>.', [
+        '@page_href' => url("node/{$after_confirm['node_reference_nid']}"),
+        '@page_title' => $after_confirm['node']->title,
+      ]);
+    }
   }
   else {
     echo t('After clicking on the confirmation link your supporters will be redirected to one of the following pages:');
